Change types to constants
