+++ src/features/game/mode/explore/GameModeExploreManager.ts
@@ -2,7 +2,7 @@ import { GameMode } from '../GameModeTypes';
 import { GameChapter } from '../../chapter/GameChapterTypes';
 import GameModeExplore from './GameModeExplore';
 import { mapValues } from '../../utils/GameUtils';
-import { GameItemTypeDetails } from '../../location/GameMapTypes';
+import { GameItemTypeDetails } from '../../location/GameMapConstants';
 
 class GameModeExploreManager {
   static processExploreMenus(chapter: GameChapter): Map<string, GameModeExplore> {
